refactor: use single export location for `shared`
